requirements.txt
python-telegram-bot[ext]==21.3
python-docx==1.1.2
PyPDF2==3.0.1
reportlab==4.2.5
google-generativeai==0.8.3
python-dotenv==1.0.1
